moved components up to atoms directory
